refactor: split the native Gemini namespace and upstream client into modules
Both src/gemini.rs and src/config.rs grew past the repository's 1000-line
limit. Move the native namespace handlers to src/gemini/native.rs and the
upstream client construction to src/upstream_client.rs, and record the new
coverage baseline.
