+++ src/lib.rs
@@ -68,6 +68,7 @@ pub mod telegram;
 pub mod token;
 pub mod token_admin;
 mod token_http;
+pub mod upstream_client;
 pub mod usage;
 pub mod vk;
 pub mod with_command;
